Set proper user agent for Hive API requests
The User-Agent header gives service providers important context
information that allows them to identify traffic and reach out to the
project if something about the API use requires changes. This, in turn,
makes it less likely that the upstream service or a third-party web
proxy like Cloudflare block the request.

Change the Hive API user agent from "Samsung Smart Fridge" to
"FlarialClient/{sha}". This agent seems less likely to be blocked in the
future than a generic client like "curl".